since bcopy/bzero will be assigned to these pointers, get the size_t in the cast right; ok millert

+++ sys/arch/mvme68k/dev/if_ie.c
@@ -1,4 +1,4 @@
-/*	$OpenBSD: if_ie.c,v 1.32 2006/01/11 07:21:58 miod Exp $ */
+/*	$OpenBSD: if_ie.c,v 1.33 2006/01/17 02:03:53 deraadt Exp $ */
 
 /*-
  * Copyright (c) 1999 Steve Murphree, Jr. 
@@ -182,9 +182,9 @@ struct ie_softc {
 	void (*reset_596)(void *);	/* card dependent reset function */
 	void (*chan_attn)(void *);	/* card dependent attn function */
 	void (*run_596)(void *);	/* card dependent "go on-line" func */
-	void (*memcopy)(const void *, void *, u_int);
+	void (*memcopy)(const void *, void *, size_t);
 	                        /* card dependent memory copy function */
-        void (*memzero)(void *, u_int);
+        void (*memzero)(void *, size_t);
 	                        /* card dependent memory zero function */
 	int want_mcsetup;       /* mcsetup flag */
 	int promisc;            /* are we in promisc mode? */
